+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 ## Added
 
+- ✨(frontend) add copy link button #235
 - 🛂(frontend) access public docs without being logged #235
 
 ## Changed